# api/openapi.yaml
openapi: 3.0.3
info:
  title: wireguardd API
  version: 0.1.0
  description: REST API for WireGuard interface and peer management.
servers:
  - url: http://127.0.0.1:51880
security:
  - bearerAuth: []
paths:
...
  /v1/interfaces/{name}/peers/{pubkey}:
    parameters:
      - $ref: "#/components/parameters/Name"
      - $ref: "#/components/parameters/PubKey"
    get:
      summary: Get peer
      responses:
        "200":
          description: Peer
    patch:
      summary: Update peer
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete peer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Name:
      name: name
      in: path
      required: true
      schema:
        type: string
    PubKey:
      name: pubkey
      in: path
      required: true
      schema:
        type: string
...